requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
